2025-03-07-Task2-incomplete
I've attempted to generate the csvs so i can move on to task 3. as you'll see from my comments in the code, i can't get it to 280. this is the version with residue 7 ('s') mutated, but the code threw up a few exceptions (sequence needs to have a phosphoacceptor) hence why i think the numbers are off. would love a bit of help with this whenever you get a chance :)

Co-Authored-By: Akashaditya Das <41631673+AkashD95@users.noreply.github.com>